requirements.txt
flask
python-dotenv
openai
streamlit
requests
gunicorn
